+++ foxplot/decode.py
@@ -6,12 +6,13 @@
 
 """Decode a series of dictionaries from file."""
 
+import json
 import sys
 from pathlib import PosixPath
 from typing import Generator, Union
 
-from .decoders.json import decode_json
-from .decoders.msgpack import decode_msgpack
+import mpacklog
+
 from .exceptions import FoxplotError
 
 
@@ -31,7 +32,32 @@ def decode(file_path: Union[str, PosixPath]) -> Generator[dict, None, None]:
         with open(file_path, "r", encoding="utf-8") as file:
             yield from decode_json(file=file)
     elif file_path.endswith(".mpack"):
-        with open(file_path, "rb") as file:
-            yield from decode_msgpack(file=file)
+        yield from mpacklog.read_log(file_path)
     else:  # unknown file extension
         raise FoxplotError(f"Unknown file type in '{file_path}'")
+
+
+def decode_json(file, chunk_size=100_000) -> Generator[dict, None, None]:
+    """Decode dictionaries from a line-delimited JSON file.
+
+    Args:
+        file: File stream (for instance ``sys.stdin``).
+        chunk_size: Number of bytes read at a time from the standard input.
+
+    Yields:
+        dict: Dictionary read from file.
+    """
+    decoder = json.JSONDecoder()
+    buffer = ""
+    while True:
+        data = file.read(chunk_size)
+        if not data:  # end of file
+            break
+        buffer += data
+        while buffer:
+            try:
+                result, index = decoder.raw_decode(buffer)
+                buffer = buffer[index:].lstrip()
+                yield result
+            except json.JSONDecodeError:
+                break